requirements.txt
numpy
csv
tqdm
matplotlib.pyplot